Search for string anywhere in line
The original version of the code captured lines starting with the input
string. However, a proposed change to the NG-CHM viewer code adds
leading whitespace for indentation to the line containing the version
number.

I tried simply changing the search string to be '  CM.version ='.
However, github actions automatically strip leading whitespace. So I
made this action search for the string in the line (e.g. `.includes()`)
rather than lines that start with the input string (e.g.
`.startsWith()`).

This change hopefully makes this action a tiny bit less brittle.